requirements.txt
openai>=1.0
whisper
ffmpeg-python
google-cloud-speech
python-docx
pydub
tqdm